:arrow_up: moves existing code into old folder in order to keep DB logic.
