+++ datastore/old/datastore.go
@@ -1,4 +1,4 @@
-package datastore
+package old
 
 import (
 	"encoding/json"
